fix: repair unread state RPC tests for imsgTests target
